OF-1517: Don't require encoded i18n files for plugins.
This commit prevents the need for \u<hex><hex><hex><hex> encoded content in translation files. This change applies the fix
that was done to the core in OF-1448 to the plugin structure.

The plugin assembly now copies all i18n files from the source folder to the build folder. There, the content is encoded
using the native2ascii maven plugin. Finally, the encoded content is used in the plugin assembly.

All plugins to which this change applies (all plugins that have i18n files) will need to get a new release. This fix requires
the new plugins to be build by Openfire 4.3.0 (which is as of yet unreleased). Runtime, there shouldn't be a need for 4.3.0.

Relevant plugins have been modified. Their version number has been increased, and the changelog modified. Their pom files have
been modified to inherit from 4.3.0-SNAPSHOT (which introduces this fix). This needs to be modified to 4.3.0 (sans -SNAPSHOT) before
the plugin can be released (we should not release code that has -SNAPSHOT dependencies). For this to be possible, Openfire 4.3.0
must be released first.